+++ vultest
@@ -15,7 +15,7 @@
 
 require 'bundler/setup'
 
-require './app'
+require './app/app'
 
 app = App.new
